Fixed an a bug which could cause the waiting writer count to overflow if another thread caused the waiting writer count reached the max, while another new writer was trying to add themselves to the count, and had already failed the CompareExchange() operation at least once.
Added a MonitorEnterTimeout property to allow the user to set the maximum amount of time that would be trying to
call Monitor.TryEnter().

Renamed the MaxWaitTime property and field so that it's clearly differentiated from the new MonitorEnterTimeout.

Modified the condition that checks if an exiting reader should release a waiting writer, so that the first condition will fail if the lock is not write reserved.  Since a writer has to reserve the lock before waiting, this will let the last exiting reader skip checking the second condition in cases where no writers are trying to enter the lock.

Added a new State and Mask constants to support the above change.

 Modified the formula to calculate the new lock state in cases where a writer exiting the lock needed to wake up a waiting writer. The new version just uses a xor to flip the WriteLockActive and ReservedForWriterFlags, since we know the write lock flag is on, and nothing can enable the write reservation flag during a write lock.

Added matching changes to the concise version of the class.